Add: Added first copies of syntax analyzer modules
Added the three modules JackCompiler, JackTokenizer, and
CompilationEngine.

JackCompiler parses the command line to get a list of .jack filenames. For
each .jack filename, JackTokenizer creates an _T.xml file which
CompilationEngine uses to create an _.xml file. This architecture is rough
right now, but it will enable me to test the essentials of the tokenizer v
ia the _T.xml file.