Mod: Replaced readlines() w/ linebyline iterations
Instead of collecting an entire file object's content in one go with
readlines(), the reading and writing from and to file objects is done one
line at a time, with 'for line in <fo>:'.  This saves computation time and
memory. The reason I was using readlines() was because I thought it would
be helpful to remove multiline comments with  regular expressions.
My plan to keep track of when multiline comments begin and end is to
set and unset a flag when /* and */ are seen.