Ref: Program structure changed
Before: JackCompiler passed a .jack filename string to JackTokenizer.
JackTokenizer opened a file object for writing to a .xml file, opened
another file object to read the .jack file, and then went through the
.jack file object line by line, writing the token classifications to the
.xml file. JackTokenizer returned a string containing the filename of the
.xml file it just wrote to. This string was passed to CompilerEngine,
which opened it for reading and wrote to another .xml file.  This
structure was thrown together to allow me to work on getting JackTokenizer
to write the correct information to an .xml file.

This commit: For each .jack file that JackCompiler is to compile,
JackCompiler opens an .xml file for JackTokenizer to write to, an .xml
file for CompilationEngine to write to, and creates a fileobject to read
the .jack file.  For each line of the .jack file object:
A. JackTokenizer is
passed the text of the line, the .xml file object it will write to, and
a flag for handling multiline comments.  JackTokenizer tokenizes the line,
writes the pertinent information to the .xml file, and returns a tuple
containing (1) a list of token information tuples and (2) the multiline
comments flag.
B. CompilationEngine is passed the list of token information tuples and
the .xml file object it will write to. All it does is write some basic
output to show that its receiving the token information it needs.

With this change, the JackCompiler works "online": it looks at .jack files
line by line and writes .xml files line by line, rather than working on an
entire file/writing an entire file/iterating any more than it needs to.