Ref: Rewrote JackTokenizer into a class
CompilationEngine is a recursive decent parser: it needs to move
recursively through the parse tree.  CompilationEngine needs to be able
to get tokens from the input while it is working recursively.  With this
consideration, JackTokenizer was rewritten into a class.  A JackTokenizer
object has the following methods that CompilationEngine can call:
tokenType()-returns the type of the current token, tokenValue()-returns
the value of the current token, advance()-advances to the next token,
returning False if there are no more tokens left.