Add: CompEng moves thru progstruct and statements
CompilationEngine moves through compilexxx() routines of the program
structure and statements varieties. There is no checking of expected/valid
tokens, and the code that is written to .xml is placeholder for the
statements and expressions compilexxx() routines.

Added writeXmlNonTerminal() and writeXmlTermina() to clean up code.

Fixed JackTokenizer so advance() doesn't advance past last token.  Bug
explanation:

WRONG - once b_condition is met, a and b both reference the same list.
(changes to b will result in instant changes to a)
a=[]
b=[]
b=update()
if(b_condition):
	a=b

CORRECT - a is set to a copy of b (changes can be made to b without
affecting a)
a=[]
b=[]
b=update()
if(b_condition):
	a=b[:]