Add: syntax verification and working compilexxx()s
Added expectTokenType() and expectTokenValue() to CompilationEngine class.
These functions accept strings separated by commas.  An error is printed
and the program is exited if the current token is not represented in the
parameter strings.

Ex: jt.TokenType() is equal to 'SYMBOL'.  If
self.expectTokenType('SYMBOL') is called, the program will move on.
Alternatively, calling
self.expectTokenType('KEYWORD','IDENTIFIER') will print an error message
to the console and exit the program.

Updated the code of the program structure and statements compilexxx()
functions.  The expectTokenXxx() functions are used to make sure the .jack
source syntax is correct (also, in combination with correct .jack source
code, the expectTokenXxx() functions were a big help in debugging JackCompiler.py).
The XML output of the parse tree is correct except for the parsing of
expressions.