Mod: Identifier xml tags include more info
With the help of the SymbolTable class, CompileClass now writes two xml lines for each identifier.  The second
line describes the category of identifer, whether it is being defined or
used, and how it is represented in the symbol table if it's a variable.
The information in these second xml lines must be known when generating VM
code.