Fix: Corrected 3 bugs in CompilationEngine
1) return for void functions and methods:
compileReturn() and compileSubroutine() now have code to push constant 0
onto the stack before the vm command return, for Jack functions that
return void.  This fix is more of a formality.  No errors were caused by
not pushing constant 0 onto the stack, because a. the vm command return
will reposition the first value on the stack to be the return value for
the caller, no matter if that value is 0 or meaningless garbage; and b.
void Jack functions are called with the keyword do, which explicitly pops
the return value from the stack and doesn't do anything with it.  Still,
it's better to return meaningful garbage that's not used, rather than
meaningless...

2) string assignment:
The last character of any string constant was not being stored when
assigning string constants to String variables.  Fixed by changing
compileTerm() to fix an off-by-1 error in the for loop that fills a String
variable by making calls to String.appendChar()

3) compileSubroutine() was defining 'this' in the symbol table too late
for method calls. 'this' is now defined in the symbol table as argument 0
before any other parameters are defined in the symbol table

Passes tests:
Seven
ConvertToBin
Square
Average
Pong