Fix: Reset after a lost ball was an instant death
Bug where ball was calculated as falling outside the play area when
starting play after a lost life

Fixed bad bug in BreakawayGame.update(), in which the hitbox of the last
brick in the wall was arbitrarily deallocated.
That memory was then being allocated to be the hitbox of the play area
when the game was reset.  However, the last brick
was still updating that memory, so the play area appeared to be a
tiny box and the ball always fell outside of it.

Edited dispose() methods of Ball, Brick, Paddle, PlayArea, and Wall to
call the dispose() method of each field before deallocating

Also in BreakawayGame
-cleaned up the processing of spacebar input in the menus
-disposed of game components after play stops, very close to where I
initialize them