Add: Testing myOS with Breakaway from project09
Testing to see if my Breakaway game from project09 will work with myOS.
It works!  But it's slow. It's slow with the SuppliedOS as well, so I can
modify the Breakaway code to try to get some speed.  Should be easier to
find gains after going through projects 10-12.

SuppliedOS is the author supplied set of .vm files for the OS classes