Mod: Added assembly functions section
Added methods to write assembly functions into an .asm file.  Added a
method to write the assembly code that will call an assembly function.

Implemented RETURN assembly function and calls to it.

Added spreadsheet to track how my changes to vmtranslator2 are affecting the number of
instructions in the translated .asm file.

Added a pause to the RUN_vmtranslator batch files.