Ref: Created VMCommandTranslator class
Created VMCommandTranslator class and integrated it into main. This
refactoring results in the exact same .asm file as before.

VMCommandTranslator contains a buffer that stores up to three vm commands.
vm commands from the input file are read into the buffer. asm is
generated based on the vm commands in the buffer.

the initial implementation of _translate() recreates the previous
one-vm-command-at-a-time translation. _translate() will be updated so that
for certain combinations of vm commands in the buffer, fewer asm
instructions will be produced for equivalent functionality