Fix: Array indexing in compileLet() bug
Say we have the following Jack statement:
let A[i] = B();

where B contains a Jack let statement with assignment to an indexed Array,
something like

B() { let C[j] = n; return x; }

My handling of the indexed Arrays in this situation led to a bug.  I was
storing the address that will be assigned to (e.g. A[i], C[j]) in a
temporary variable to be held while the expression on the right side of
the assignment was being evaluated. I was using the same temporary
variable for every let statement.  So in the example above, storing C[j]
would blow away the stored A[i], and there was no recovery from this.

I fixed this by leaving A[i]/C[j] on the stack while the expression on the
right side of the assignment is evaluated.  The result of the expression
is then placed upon the stack.  This result is popped from the stack and stored in a temporary variable
while the A[i]/C[j] value on the stack is used to set up the pop to the
address pointed to by A[i]/C[j].  The result in the temporary variable is
returned to the stack, and then popped to the correct address that was
pointed to by A[i]/C[j].